requirements.txt
pandas
psycopg2-binary
python-dotenv
gspread
gspread-dataframe
oauth2client
SQLAlchemy
